Add info to runbook on Prom disk space filling
This will mean that all EBS volumes will be increased.
The reason we are doing this is because if we increase just one EBS volume outside of terraform, this will cause  a problem with applying terraform as it will attempt to decrease the EBS volume size to the terraform size.