Change `max_toc_heading_level` from 6 to 3
In a recent PR https://github.com/alphagov/re-team-manual/pull/53
I added further level headings to the alerts section of the
observe playbook. As the `max_toc_heading_level` is 6 it means that
all of these headings are added to the navigation. I don't think
this is particularly useful and clutters up the navigation menu.

I have reduced the number down to 3 and I think this gives a
useable menu whilst still providing enough depth to our navigation
so users can find what they want.

This hasn't been tested with users but is just based on what
I think looks sensible.