Document the use of YubiKeys
The GSP team started to enrol some YubiKeys to a service team. They have
come up with a set of instructions to make it as neat as possible for
the end user to set that up.

This should come in handy in the near future as the secret evil plan is
to start a GDS wide spread of these.

Co-authored-by: Rafal Proszowski <rafal.proszowski@digital.cabinet-office.gov.uk>
Co-authored-by: Chris Farmiloe <chris.farmiloe@digital.cabinet-office.gov.uk>
Co-authored-by: Anshul Sirur <anshul.sirur@digital.cabinet-office.gov.uk>
Co-authored-by: Daniel Blair <daniel.blair@digital.cabinet-office.gov.uk>